+++ packages/topofit/CHANGELOG.md
@@ -5,6 +5,12 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @neurodesk/webapp-components@0.4.1
+
+
+### Patch Changes
+
+- Updated dependencies
   - @neurodesk/webapp-components@0.4.0
 
 ## 0.7.20260917